9.Bölüm ödevindeki veri tipi sorunu düzeltildi.

+++ Entities/Concrete/Car.cs
@@ -9,7 +9,7 @@ namespace Entities.Concrete
         public int BrandId { get; set; } //Marka Id
         public int ColorId { get; set; } //Renk Id
         public int ModelYear { get; set; }
-        public int DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
         public string Description { get; set; }
     }
 }
